requirements.txt
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.17
langchain-google-genai
python-dotenv==1.0.1
streamlit==1.30.0
chromadb==0.4.24
beautifulsoup4==4.12.2
sentence-transformers
requests